+++ crates/cli/tests/check_test.rs
@@ -55,3 +55,32 @@ fn runs_tasks_from_multiple_project() {
     assert!(predicate::str::contains("noop:noopWithDeps").eval(&output));
     assert!(predicate::str::contains("depsA:dependencyOrder").eval(&output)); // dep of noop
 }
+
+mod reports {
+    use super::*;
+
+    #[test]
+    fn doesnt_create_a_report_by_default() {
+        let fixture = create_sandbox_with_git("cases");
+
+        create_moon_command(fixture.path())
+            .arg("check")
+            .arg("base")
+            .assert();
+
+        assert!(!fixture.path().join(".moon/cache/runReport.json").exists());
+    }
+
+    #[test]
+    fn creates_report_when_option_passed() {
+        let fixture = create_sandbox_with_git("cases");
+
+        create_moon_command(fixture.path())
+            .arg("check")
+            .arg("base")
+            .arg("--report")
+            .assert();
+
+        assert!(fixture.path().join(".moon/cache/runReport.json").exists());
+    }
+}